Refactoring how priors work... not done yet
